+++ R/gm.measurement.error.R
@@ -34,7 +34,10 @@
 #' (as an interaction with replicates)..
 #' This would be of interest if one were concerned with systematic ME occurring perhaps differently among 
 #' certain strata within the data.  For example, systematic ME because of an observer bias might
-#' only be observed with females or males.  
+#' only be observed with females or males. 
+#' @param groups.first A logical value for whether to use groups as a term before subjects, so that it can be included
+#' in an ANOVA table (if TRUE).  Otherwise, a group effect is likely subsumed by a subject effect, since subjects are 
+#' unique to groups. 
 #' @param data An data frame of class \code{\link{geomorph.data.frame}}.
 #' @param iter Number of iterations for significance testing
 #' @param seed An optional argument for setting the seed for random 
@@ -136,6 +139,7 @@ gm.measurement.error <- function(coords,
                                  subjects, 
                                  replicates, 
                                  groups = NULL,
+                                 groups.first = FALSE,
                                  data,
                                  iter = 999, 
                                  seed = NULL,
@@ -178,6 +182,7 @@ gm.measurement.error <- function(coords,
     subjects = subjects, 
     replicates = replicates, 
     groups = groups,
+    groups.first = groups.first,
     data = data,
     iter = iter, 
     seed = seed,