+++ NEWS.md
@@ -1,7 +1,13 @@
 # CHANGES IN GEOMORPH VERSION 4.0.8.999
 
+### NEW FEATURES
+
+* New function `physignal.eigen` for evaluating the dimensions of phylogenetic signal in multivariate data
+* Added an option to `gm.prcomp` to allow the phylogeny to be scaled to unit length
+
 ### BUG FIXES
-+ Fixed the missing re-order of data for `anc.BM`.
+* Ensured proper size-scaling of predicted shapes in `gm.prcomp`
+* Fixed the missing re-order of data for `anc.BM`.
 
 # CHANGES IN GEOMORPH VERSION 4.0.8